requirements.txt
discord.py==2.3.1
mcipc==2.4.2
mojang==1.1.0
python-dotenv==1.0.0
rcon-version==2.3.9
virtualenv==20.24.0